fix: extract SmallHeader into its own client component file
Header.tsx having "use client" caused SiteHeader (a server component
with no "use client") to not SSR, breaking the header. Isolate the
client boundary to SmallHeader.tsx only.
